Use TimeoutSource directly instead of Timeout
(In Playback for progress tracking)

This allows us to check if source is destroyed rather than
keeping an id of the source and checking if it is zero or not.

Otherwise, When the source is removed via something other than
stop_progress_tracking() e.g. the callback (update_progress())
returning false or Source.REMOVE, our timeout_id variable does
not automatically reset to zero and causes stop_progress_tracking()
to print errers on the terminal when it is called.