Animate video resize when turning on "Overlay UI"
Video resize when turning "Overlay UI" off was already
animated because of the Revealers' animated resize coupled
with Box-like layout of the top/bottom bars and video.

On the other hand, turning on overlay ui changed the layout
to an overlay layout instantly (just like turning it off
changed the layout to a Box-like layout instantly). But we
need a Box-like layout for the video to resize with
resizing/animated Revealers i.e. animate alongside Revealers.

That is why this commits makes it so that we wait for
overlay_ui_turning_on animation (tracked via child-revealed
property) to complete before switching to an overlay layout.