MPRIS: Expose filename in place of empty metadata
Previously, when the title or any other metadata of a media stream
was unkown, we were still exposing those fields with empty string
values, which resulted in broken player control UI of the system
(in case of GNOME at least).

Now, when the title is unknown, we expose filename as the title,
and when any other metadata is unkown, we simply don't expose
the associated MPRIS fields.